Major functionality overhaul.
Here we take a more graceful approach to injecting functionality around
table inheritance into ActiveRecord.

First, we check once on boot to
gather the necessary information from the DB around parent/child table
relationships. This is cached for use later at runtime to eliminate
unnecessary calls to the DB to constantly check inheritance status. It's
safe to assume table inheritance does not change at runtime, so we
can afford to check this once.

Second, the logic around when and how to inject functionality into the
appropriate ActiveRecord::Model has been optimized. This takes advantage
of some public and private APIs in ActiveRecord to more seamlessly
detect when the right time to engage is.

Finally, as with the improvements for functionality injection, the logic
around modifying the select statements

As a result of these optimization efforts, support for ActiveRecord less
than 4.2 has been dropped.

Additionally, partially implemented features such as some
reconfigurability has been removed until it can be implemented fully.